refactor: use key in instead of instanceof
key in is faster in most browsers https://jsbench.me/3qkrn7gwez/2
