+++ packages/reactivity/__tests__/deferredComputed.spec.ts
@@ -1,12 +1,4 @@
-import {
-  computed,
-  deferredComputed,
-  effect,
-  isComputed,
-  reactive,
-  ref,
-  toRaw
-} from '../src'
+import { computed, deferredComputed, effect, ref } from '../src'
 
 describe('deferred computed', () => {
   const tick = Promise.resolve()
@@ -189,14 +181,5 @@ describe('deferred computed', () => {
     src.value++
     await tick
     expect(c1Spy).toHaveBeenCalledTimes(1)
-  })
-
-  it('isComputed', () => {
-    expect(isComputed(computed(() => 1))).toBe(true)
-    expect(isComputed(deferredComputed(() => 1))).toBe(true)
-    expect(isComputed(ref(1))).toBe(false)
-    expect(isComputed(reactive({}))).toBe(false)
-    expect(isComputed(reactive({ c: computed(() => 1) }).c)).toBe(false)
-    expect(isComputed(toRaw(reactive({ c: computed(() => 1) })).c)).toBe(true)
   })
 })
