+++ packages/reactivity/src/deferredComputed.ts
@@ -1,6 +1,6 @@
 import { Dep } from './dep'
 import { ReactiveEffect } from './effect'
-import { ComputedGetter, ComputedRef, ComputedRefImpl } from './computed'
+import { ComputedGetter, ComputedRef } from './computed'
 import { ReactiveFlags, toRaw } from './reactive'
 import { trackRefValue, triggerRefValue } from './ref'
 
@@ -22,15 +22,6 @@ const flush = () => {
   }
   queue.length = 0
   queued = false
-}
-
-export function isComputed<T>(
-  value: ComputedRef<T> | unknown
-): value is ComputedRef<T>
-export function isComputed(value: any): value is ComputedRef {
-  return (
-    value instanceof ComputedRefImpl || value instanceof DeferredComputedRefImpl
-  )
 }
 
 class DeferredComputedRefImpl<T> {
